Create FacebookNews model

+++ app/src/main/java/com/example/robert/medius/newsFeed/NewsFeedInteractorImpl.kt
@@ -21,7 +21,7 @@ class NewsFeedInteractorImpl(val twitterApiHelper: TwitterApiHelper, val faceboo
     override fun loadMoreTimeline(lastItem: News?, feedType: NewsFeedType) {
         when (feedType) {
             NewsFeedType.FACEBOOK -> facebookApiHelper.loadMoreTimeline()
-            NewsFeedType.TWITTER -> twitterApiHelper.loadMoreTimeline(if (lastItem != null) lastItem.id - 1 else null)
+            NewsFeedType.TWITTER -> twitterApiHelper.loadMoreTimeline(if (lastItem != null) lastItem.id.toLong() - 1 else null)
             NewsFeedType.NONE -> throw IllegalStateException("Calling this method when a NewsFeedType of the fragment is equals to NONE is not allowed")
         }
     }
